requirements.etl.txt
psycopg2-binary==2.9.9
elasticsearch==8.6.2
